+++ src/about.tcl
@@ -89,6 +89,8 @@ proc about::About {} {
     [list "link_" "::apave::openDoc %t@@https://aplsimple.github.io/en/misc/links/links.html@@"] \
     [list "linkRH" "::apave::openDoc %t@@http://www.hwaci.com/drh/@@"] \
     [list "linkBL" "::apave::openDoc %t@@https://wiki.tcl-lang.org/page/bll@@"] \
+    [list "linkFF" "::apave::openDoc %t@@https://wiki.tcl-lang.org/page/FF@@"] \
+    [list "linkSS" "::apave::openDoc %t@@https://github.com/antirez@@"] \
     ]
 
   ### ________________________ "General" tab _________________________ ###
@@ -152,6 +154,8 @@ proc about::About {} {
     \u2022 <linkCW>Christian Werner</linkCW>\n \
     \u2022 <linkNB>Nicolas Bats</linkNB>\n \
     \u2022 <linkTZ>Thanos Zygouris</linkTZ>\n \
+    \u2022 <linkFF>Federico Ferri</linkFF>\n \
+    \u2022 <linkSS>Salvatore Sanfilippo</linkSS>\n \
     \u2022 Alexis Martin\n \
     \n $spec\n\n \
     \u2022 <linkTW>Trevor Williams</linkTW>\n \