+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # LAST CHANGES:
 
 
-Version `1.3.5.2 (21 Dec'22)`
+Version `1.3.6a2 (24 Dec'22)`
 
   - BUGFIX: handling keys: Tab and Shift+Tab on texts, esp. in Windows
   - BUGFIX: -defaultextension .tcl at saving files: for "no name" only
+  - BUGFIX: skip Tcl check at Run if unit tree not built yet
+  - NEW   : 3 profilers, available via Templates
+  - NEW   : save/restore Templates' selection
+  - NEW   : %A and %M wildcards for Templates
   - CHANGE: syntax highlighting in Templates
   - CHANGE: hot keys in tooltips/menu of file tree
   - CHANGE: clear the unit error message in info bar, before Run
@@ -12,7 +16,7 @@ Version `1.3.5.2 (21 Dec'22)`
   - CHANGE: tips on Projects' file list
   - CHANGE: About / Acknowledgements
   - CHANGE: docs
-  - CHANGE: packages: apave/e_menu 3.6.1
+  - CHANGE: packages: apave/e_menu 3.6.2
 
 
 Version `1.3.5 (7 Dec'22)`